Refactor to common for AB#10656
