+++ Apps/Common/test/unit/CommonTests.xml
@@ -416,6 +416,46 @@
             <param name="throwException">Throw exception indicator.</param>
             <returns>Mocked notification settings.</returns>
         </member>
+        <member name="T:HealthGateway.CommonTests.Delegates.WalletIssuerDelegateTests">
+            <summary>
+            Unit Tests for the Wallet Issuer Delegate.
+            </summary>
+        </member>
+        <member name="M:HealthGateway.CommonTests.Delegates.WalletIssuerDelegateTests.#ctor">
+            <summary>
+            Initializes a new instance of the <see cref="T:HealthGateway.CommonTests.Delegates.WalletIssuerDelegateTests"/> class.
+            </summary>
+        </member>
+        <member name="M:HealthGateway.CommonTests.Delegates.WalletIssuerDelegateTests.CreateConnection200">
+            <summary>
+            Create Connection - Happy Path.
+            </summary>
+        </member>
+        <member name="M:HealthGateway.CommonTests.Delegates.WalletIssuerDelegateTests.CreateConnection500">
+            <summary>
+            Create Connection - Agent down.
+            </summary>
+        </member>
+        <member name="M:HealthGateway.CommonTests.Delegates.WalletIssuerDelegateTests.CreateConnectionBadPayload">
+            <summary>
+            Create Connection - Bad Payload from Agent.
+            </summary>
+        </member>
+        <member name="M:HealthGateway.CommonTests.Delegates.WalletIssuerDelegateTests.CreateConnectionException">
+            <summary>
+            Create Connection - Exception during HTTP Call.
+            </summary>
+        </member>
+        <member name="M:HealthGateway.CommonTests.Delegates.WalletIssuerDelegateTests.CreateConnection(System.Net.HttpStatusCode,HealthGateway.Common.Models.RequestResult{HealthGateway.Common.Models.AcaPy.ConnectionResponse},System.Boolean,System.Boolean)">
+            <summary>
+            Create Wallet Connection.
+            </summary>
+            <param name="expectedResponseStatusCode">expectedResponseStatusCode.</param>
+            <param name="expectedRequestResult">expectedRequestResult.</param>
+            <param name="badContent">Pass no json data as as response.</param>
+            <param name="throwException">Throw exception indicator.</param>
+            <returns>The notification settings.</returns>
+        </member>
         <member name="T:HealthGateway.CommonTests.FileDownload.FileDownloadServiceTests">
             <summary>
             FileDownloadService's Unit Tests.
